fix: 참여한 기부 리스트 API 정상 동작하도록 수정

+++ src/main/java/com/project/give/dto/donation/response/GetDonationProjectsResponseDto.java
@@ -25,6 +25,6 @@ public class GetDonationProjectsResponseDto {
     private LocalDateTime createDate;
     private LocalDateTime updateDate;
     private String donationCategoryNameKor;
-    private BigDecimal totalContribution;
+    private int totalContribution;
     private LocalDateTime lastContributionDate;
 }
